feat: configurable up binary path
The extension was crashing at launch because of it not being able to
find the "up" binary in the path, making it configurable and setting it
to the correct value fixed the issue. The configuration change requires
a manual reload of the vscode window, which can be done from the
command palette, "Developer: Reload Window". In the future we could
think of implementing automatic restart on this parameter change.

Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>